+++ source/codegen/metadata/nitclk/config.py
@@ -1,0 +1,39 @@
+# -*- coding: utf-8 -*-
+# This file is generated from NI-TClk API metadata version 255.0.0d0
+config = {
+    'api_version': '255.0.0d0',
+    'c_header': 'niTClk.h',
+    'c_function_prefix': 'niTClk_',
+    'service_class_prefix': 'NiTClk',
+    'java_package': 'com.ni.grpc.tclk',
+    'csharp_namespace': 'NationalInstruments.Grpc.TClk',
+    'namespace_component': 'nitclk',
+    'close_function': None,
+    'custom_types': [
+    ],
+    'driver_name': 'NI-TClk',
+    'extra_errors_used': [],
+    'library_info': {
+        'Linux': {
+            '64bit': {
+                'name': 'nitclk',
+                'type': 'cdll'
+            }
+        },
+        'Windows': {
+            '32bit': {
+                'name': 'nitclk.dll',
+                'type': 'windll'
+            },
+            '64bit': {
+                'name': 'nitclk_64.dll',
+                'type': 'cdll'
+            }
+        }
+    },
+    'metadata_version': '2.0',
+    'module_name': 'nitclk',
+    'repeated_capabilities': [
+    ],
+    'session_handle_parameter_name': 'session_number',
+}