+++ source/server/core_server.cpp
@@ -1,6 +1,8 @@
 #include <mutex>
 #include <nidcpower/nidcpower_library.h>
 #include <nidcpower/nidcpower_service.h>
+#include <nidigitalpattern/nidigitalpattern_library.h>
+#include <nidigitalpattern/nidigitalpattern_service.h>
 #include <nidmm/nidmm_library.h>
 #include <nidmm/nidmm_service.h>
 #include <niscope/niscope_library.h>
@@ -9,6 +11,8 @@
 #include <niswitch/niswitch_service.h>
 #include <nisync/nisync_library.h>
 #include <nisync/nisync_service.h>
+#include <nitclk/nitclk_library.h>
+#include <nitclk/nitclk_service.h>
 
 #include "logging.h"
 #include "server_configuration_parser.h"
@@ -79,6 +83,10 @@ static void RunServer(const ServerConfiguration& config)
   nidevice_grpc::SessionUtilitiesService core_service(&session_repository, &device_enumerator);
   builder.RegisterService(&core_service);
 
+  nidigitalpattern_grpc::NiDigitalLibrary nidigital_library;
+  nidigitalpattern_grpc::NiDigitalService nidigital_service(&nidigital_library, &session_repository);
+  builder.RegisterService(&nidigital_service);
+
   niscope_grpc::NiScopeLibrary niscope_library;
   niscope_grpc::NiScopeService niscope_service(&niscope_library, &session_repository);
   builder.RegisterService(&niscope_service);
@@ -95,6 +103,10 @@ static void RunServer(const ServerConfiguration& config)
   nisync_grpc::NiSyncService nisync_service(&nisync_library, &session_repository);
   builder.RegisterService(&nisync_service);
 
+  nitclk_grpc::NiTClkLibrary nitclk_library;
+  nitclk_grpc::NiTClkService nitclk_service(&nitclk_library, &session_repository);
+  builder.RegisterService(&nitclk_service);
+  
   nidcpower_grpc::NiDCPowerLibrary nidcpower_library;
   nidcpower_grpc::NiDCPowerService nidcpower_service(&nidcpower_library, &session_repository);
   builder.RegisterService(&nidcpower_service);